Working with random numbers.

+++ MonteCarlo/MonteCarloResult.cs
@@ -6,7 +6,7 @@ namespace SmartMaintenance.MonteCarlo
     {
         // Public properties
         public int FoundAtLoop { get; set; }
-        public decimal AdjustedReliability { get; set; }
+        public double AdjustedReliability { get; set; }
         public VariableInput[] Inputs { get; set; }
     }
 }
